* audio/midi/* : Changes to the MidiDriver code as per Exult bug report 	  1750743, with additional changes to ensure that it actually works 	  properly in all cases. Also fixed possible problems with sequence speed 	  being set to 0, notes being cut off early at the end of a song and 	  renamed some variables so they properly reflect what they actually are. 	* world/GravityProcess.cpp : Removing ambiguous call for MSVC 	* system/msvc/msvc_include.h : Adding an define required so pentagram 	  will compile 	* system/msvc/Pentagram/Pentagram.vcproj : Updating project to include all 	  files

+++ world/GravityProcess.cpp
@@ -188,7 +188,7 @@ bool GravityProcess::run(uint32 framenum)
 
 				// Apply an impulse on the x/y plane in a random direction
 				// in a 180 degree pie around the orginal vector in x/y
-				double heading_r = atan2(yspeed, xspeed);
+				double heading_r = atan2((double)yspeed, (double)xspeed);
 				double deltah_r = static_cast<double>(rand())
 								  * M_PI / RAND_MAX - M_PI/2;
 #ifdef BOUNCE_DIAG
